move erasure to service/ dir
